Add FORTIFY_SOURCE fix and pin package versions
As described in #77 and #79, FORTIFY_SOURCE is needed for Cmake to work
on icamerasrc and intel-ipu36-camera-hal. Additionally, recent upstream
releases have broken several dependencies. While I'm hoping the recent
flurry of activity upstream means support for kernel 6.7+ is on the way,
these workarounds should keep us running for now.